+++ hls/src/krnl_vadd.h
@@ -11,5 +11,5 @@ static constexpr uint64_t MAX_COLS = 512;//687;
 static constexpr uint64_t MAX_PIXELS = MAX_COLS * MAX_ROWS;
 
 void krnl_vadd(Pixel *src_ptr, Pixel *dst_ptr, uint16_t rows, uint16_t cols,
-		uint8_t direction);
+		uint16_t direction);
